refactor: extract shared MCP client helpers into gateway_mcp_common
