+++ packages/gateway_mcp_common/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "gateway_mcp_common"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_mcp_common"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
