+++ packages/gateway_notes_mcp/pyproject.toml
@@ -2,7 +2,7 @@
 name = "gateway_notes_mcp"
 version = "3.1.0"
 requires-python = ">=3.11"
-dependencies = ["gateway_core", "gateway_notes", "fastmcp>=3.4.2"]
+dependencies = ["gateway_core", "gateway_notes", "gateway_mcp_common", "fastmcp>=3.4.2"]
 
 [project.entry-points."substack_gateway.capabilities"]
 notes_mcp = "gateway_notes_mcp:capability"
@@ -17,3 +17,4 @@ packages = ["src/gateway_notes_mcp"]
 [tool.uv.sources]
 gateway_core = { workspace = true }
 gateway_notes = { workspace = true }
+gateway_mcp_common = { workspace = true }
