+++ packages/gateway_oss/pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
   "gateway_notes_rest",
   "gateway_notes_mcp",
   "gateway_rest_common",
+  "gateway_mcp_common",
 ]
 
 [build-system]
@@ -33,3 +34,4 @@ gateway_notes = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_rest_common = { workspace = true }
+gateway_mcp_common = { workspace = true }
